Update 2019_07_24_083500_create_products_table.php
fix

+++ database/migrations/2019_07_24_083500_create_products_table.php
@@ -15,12 +15,12 @@ class CreateProductsTable extends Migration
     {
         Schema::create('products', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('brand_id')->index();
+            $table->unsignedBigInteger('brand_id')->index();
             $table->string('sku');
             $table->string('name');
             $table->string('slug');
             $table->text('description')->nullable();
-            $table->unsignedInteger('quantity');
+            $table->unsignedBigInteger('quantity');
             $table->decimal('weight', 8, 2)->nullable();
             $table->decimal('price', 8, 2)->nullable();
             $table->decimal('sale_price', 8, 2)->nullable();
